cleanup current Sandboxing code
