+++ src/core/src/main/scala/gwi/mawex/Executor.scala
@@ -2,6 +2,7 @@ package gwi.mawex
 
 import akka.actor.SupervisorStrategy.{Escalate, Stop}
 import akka.actor._
+import gwi.mawex.k8.K8SandBox
 
 /** Executor itself is delivered by users, Mawex provides only SandBox where Executor runs, see [[gwi.mawex.WorkerCmd]] */
 object Executor {
@@ -22,16 +23,16 @@ trait SandBox extends Actor with ActorLogging {
 
 object SandBox {
   val ActorName = "SandBox"
-  def localJvmProps(executorProps: Props): Props = Props(classOf[LocalJvmSandBox], executorProps)
+  def localJvmProps(executorProps: Props): Props = Props(classOf[LocalJvmExecutor], executorProps)
   def forkingProps(executorProps: Props, forkedJvm: ForkedJvm): Props = Props(classOf[ForkingSandBox], executorProps, forkedJvm)
-  def k8JobProps(executorProps: Props, forkedJvm: ForkedJvm): Props = Props(classOf[ForkingSandBox], executorProps, forkedJvm)
+  def k8JobProps(executorProps: Props, forkedJvm: ForkedJvm): Props = Props(classOf[K8SandBox], executorProps, forkedJvm)
 }
 
 /** SandBox for local JVM execution */
-class LocalJvmSandBox(executorProps: Props) extends SandBox {
+class LocalJvmExecutor(props: Props) extends SandBox {
   override def receive: Receive = {
     case Task(_, job) =>
-     val executor = context.child(Executor.ActorName) getOrElse context.actorOf(executorProps, Executor.ActorName)
+     val executor = context.child(Executor.ActorName) getOrElse context.actorOf(props, Executor.ActorName)
       executor.forward(job)
     }
 }
