+++ src/core/src/main/scala/gwi/mawex/ForkingSandBox.scala
@@ -1,13 +1,11 @@
 package gwi.mawex
 
-import java.util.concurrent.atomic.AtomicBoolean
-import akka.actor.{Actor, ActorRef, ActorSelection, Address, AddressFromURIString, Deploy, PoisonPill, Props, ReceiveTimeout}
+import akka.actor.{ActorRef, Deploy, PoisonPill, Props, ReceiveTimeout}
 import akka.remote.{DisassociatedEvent, RemoteScope}
 import akka.serialization.Serialization
-import org.backuity.clist.{CliMain, arg}
-import scala.concurrent.duration._
-import scala.concurrent.{Await, ExecutionContext}
-import scala.concurrent.duration.Duration
+import org.backuity.clist.CliMain
+
+import scala.concurrent.duration.{Duration, _}
 import scala.sys.process.Process
 import scala.util.Try
 
@@ -19,10 +17,10 @@ case class ForkedJvm(classPath: String, opts: String)
 class ForkingSandBox(executorProps: Props, forkedJvm: ForkedJvm) extends SandBox {
 
   private[this] var process: Option[Process] = Option.empty
-  private[this] var frontDesk: Option[ActorRef] = Option.empty
+  private[this] var frontDeskRef: Option[ActorRef] = Option.empty
 
   private[this] def terminateProcess() = {
-    frontDesk.foreach(_ ! s2e.TerminateExecutor)
+    frontDeskRef.foreach(_ ! s2e.TerminateExecutor)
     (1 to 10).foldLeft(process.exists(_.isAlive())) {
       case (acc, counter) if acc =>
         Thread.sleep(500)
@@ -32,7 +30,7 @@ class ForkingSandBox(executorProps: Props, forkedJvm: ForkedJvm) extends SandBox
         false
     }
     process = Option.empty
-    frontDesk = Option.empty
+    frontDeskRef = Option.empty
   }
 
   override def preStart(): Unit = context.system.eventStream.subscribe(self, classOf[DisassociatedEvent])
@@ -51,7 +49,7 @@ class ForkingSandBox(executorProps: Props, forkedJvm: ForkedJvm) extends SandBox
     case e2s.RegisterExecutor =>
       val address = sender().path.address
       log.info(s"Forked Executor registered at $address")
-      frontDesk = Some(sender())
+      frontDeskRef = Some(sender())
       context.setReceiveTimeout(Duration.Undefined)
       val executor = context.actorOf(executorProps.withDeploy(Deploy(scope = RemoteScope(address))), Executor.ActorName)
       executor ! job
@@ -76,30 +74,9 @@ class ForkingSandBox(executorProps: Props, forkedJvm: ForkedJvm) extends SandBox
   }
 }
 
-object ForkingSandBox extends CliMain[Unit]() {
-  class ExecutorDoorman(sandbox: ActorSelection) extends Actor {
-    override def preStart(): Unit = sandbox ! e2s.RegisterExecutor
-    override def receive: Receive = {
-      case s2e.TerminateExecutor => context.system.terminate()
-    }
-  }
+object ForkingSandBox extends CliMain[Unit]() with SandBoxLauncherSupport {
 
-  var sandboxActorPath = arg[String](name="sandbox-actor-path", description = "Serialization.serializedActorPath")
-
-  private val systemTerminated = new AtomicBoolean(false)
-
-  override def run: Unit = {
-    val sandboxAddress = AddressFromURIString(sandboxActorPath)
-    val executorAddress = Address("akka.tcp", Executor.SystemName, sandboxAddress.host.get, 0)
-    val executorSystem = RemoteService.buildRemoteSystem(executorAddress)
-    val sandbox = executorSystem.actorSelection(sandboxActorPath)
-    executorSystem.actorOf(Props(classOf[ExecutorDoorman], sandbox))
-    executorSystem.whenTerminated.onComplete { _ =>
-      systemTerminated.set(true)
-      System.exit(0)
-    }(ExecutionContext.Implicits.global)
-    sys.addShutdownHook(if (!systemTerminated.get) Await.result(executorSystem.terminate(), 10.seconds))
-  }
+  override def run: Unit = startAndRegisterExecutorToSandBox
 
   def fork(sandBox: ActorRef, forkedJvm: ForkedJvm): Process =
     Fork.run(
